+++ src/Util/IntrospectionProcessor.php
@@ -16,6 +16,7 @@ use Shopware\Core\Framework\HttpException;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\ShopwareHttpException;
 use Swag\PayPal\AgentCommerce\Exception\AgentHttpException;
+use Shopware\Core\Framework\Validation\Exception\ConstraintViolationException;
 use Swag\PayPal\Pos\Api\Exception\PosException;
 use Swag\PayPal\Pos\Client\AbstractClient as PosAbstractClient;
 use Swag\PayPal\RestApi\Client\AbstractClient;
@@ -145,12 +146,7 @@ class IntrospectionProcessor implements ProcessorInterface
      */
     private function exceptionToContext(\Throwable $exception): array
     {
-        $context = [
-            'message' => $exception->getMessage(),
-            'class' => $this->traceToClassString($exception->getTrace()[0]),
-            'file' => $exception->getFile(),
-            'line' => $exception->getLine(),
-        ];
+        $context = ['message' => $exception->getMessage()];
 
         if ($exception instanceof AgentHttpException) {
             $context['details'] = $exception->getDetails()->map(static fn (PayPalApiStruct $e): array => $e->jsonSerialize());
@@ -160,9 +156,20 @@ class IntrospectionProcessor implements ProcessorInterface
             $context['parameters'] = $exception->getParameters();
         }
 
+        if ($exception instanceof ConstraintViolationException) {
+            foreach ($exception->getViolations() as $violation) {
+                $context['parameters']['violations'][] = (string) $violation;
+            }
+        }
+
         if ($exception instanceof HttpException || $exception instanceof PosException) {
             $context['errorCode'] = $exception->getErrorCode();
         }
+
+        // Order class, file and line at the end to make the exception & the most important information more readable in logs
+        $context['class'] = $this->traceToClassString($exception->getTrace()[0]);
+        $context['file'] = $exception->getFile();
+        $context['line'] = $exception->getLine();
 
         if ($exception->getPrevious()) {
             $context['previous'] = $this->exceptionToContext($exception->getPrevious());